strange moves with strings
