+++ src/UserService.Mappers/Models/Interfaces/ICertificateInfoMapper.cs
@@ -1,7 +1,6 @@
 ﻿using LT.DigitalOffice.Kernel.Attributes;
 using LT.DigitalOffice.UserService.Models.Db;
 using LT.DigitalOffice.UserService.Models.Dto.Models;
-using LT.DigitalOffice.UserService.Models.Dto.Models.Certificates;
 
 namespace LT.DigitalOffice.UserService.Mappers.Models.Interfaces
 {
