+++ src/UserService.Mappers/Models/UserInfoMapper.cs
@@ -26,14 +26,14 @@ namespace LT.DigitalOffice.UserService.Mappers.Models
                 FirstName = dbUser.FirstName,
                 LastName = dbUser.LastName,
                 MiddleName = dbUser.MiddleName,
-                Gender = (UserGender)dbUser.Gender,
+                Gender = ((UserGender)dbUser.Gender).ToString(),
                 DateOfBirth = dbUser.DateOfBirth?.ToShortDateString(),
                 City = dbUser.City,
                 StartWorkingAt = dbUser.StartWorkingAt?.ToShortDateString(),
                 About = dbUser.About,
                 IsAdmin = dbUser.IsAdmin,
                 IsActive = dbUser.IsActive,
-                Status = (UserStatus)dbUser.Status,
+                Status = ((UserStatus)dbUser.Status).ToString(),
                 Rate = dbUser.Rate,
                 Office = office,
                 Role = role,
