+++ src/UserService.Models.Dto/Models/EducationInfo.cs
@@ -1,5 +1,4 @@
-﻿using LT.DigitalOffice.UserService.Models.Dto.Enums;
-using System;
+﻿using System;
 
 namespace LT.DigitalOffice.UserService.Models.Dto.Models
 {
@@ -8,7 +7,7 @@ namespace LT.DigitalOffice.UserService.Models.Dto.Models
         public Guid Id { get; set; }
         public string UniversityName { get; set; }
         public string QualificationName { get; set; }
-        public FormEducation FormEducation { get; set; }
+        public string FormEducation { get; set; }
         public DateTime AdmissionAt { get; set; }
         public DateTime? IssueAt { get; set; }
     }
