+++ src/UserService.Models.Dto/Models/UserInfo.cs
@@ -1,5 +1,4 @@
-﻿using LT.DigitalOffice.UserService.Models.Dto.Enums;
-using System;
+﻿using System;
 
 namespace LT.DigitalOffice.UserService.Models.Dto.Models
 {
@@ -9,10 +8,10 @@ namespace LT.DigitalOffice.UserService.Models.Dto.Models
         public string FirstName { get; set; }
         public string LastName { get; set; }
         public string MiddleName { get; set; }
-        public UserGender Gender { get; set; }
+        public string Gender { get; set; }
         public string DateOfBirth { get; set; }
         public string City { get; set; }
-        public UserStatus Status { get; set; }
+        public string Status { get; set; }
         public bool IsAdmin { get; set; }
         public string About { get; set; }
         public string StartWorkingAt { get; set; }
