+++ src/UserService.Models.Dto/Responses/User/UserResponse.cs
@@ -1,5 +1,4 @@
 ﻿using LT.DigitalOffice.UserService.Models.Dto.Models;
-using LT.DigitalOffice.UserService.Models.Dto.Models.Certificates;
 using System.Collections.Generic;
 
 namespace LT.DigitalOffice.UserService.Models.Dto.Responses.User
