+++ test/UserService.Mappers.UnitTests/Responses/UserResponseMapperTests.cs
@@ -5,7 +5,6 @@ using LT.DigitalOffice.UserService.Models.Db;
 using LT.DigitalOffice.UserService.Models.Dto;
 using LT.DigitalOffice.UserService.Models.Dto.Enums;
 using LT.DigitalOffice.UserService.Models.Dto.Models;
-using LT.DigitalOffice.UserService.Models.Dto.Models.Certificates;
 using LT.DigitalOffice.UserService.Models.Dto.Requests.User.Filters;
 using LT.DigitalOffice.UserService.Models.Dto.Responses.User;
 using Moq;
@@ -150,7 +149,7 @@ namespace LT.DigitalOffice.UserService.Mappers.ResponsesMappers.UnitTests
                 Id = _dbUserEducation.Id,
                 UniversityName = _dbUserEducation.UniversityName,
                 QualificationName = _dbUserEducation.QualificationName,
-                FormEducation = (FormEducation)_dbUserEducation.FormEducation,
+                FormEducation = _dbUserEducation.FormEducation.ToString(),
                 AdmissionAt = _dbUserEducation.AdmissionAt,
                 IssueAt = _dbUserEducation.IssueAt
             };
@@ -210,10 +209,10 @@ namespace LT.DigitalOffice.UserService.Mappers.ResponsesMappers.UnitTests
                 FirstName = _dbUser.FirstName,
                 MiddleName = _dbUser.MiddleName,
                 DateOfBirth = _dbUser.DateOfBirth.ToString(),
-                Gender = (UserGender)_dbUser.Gender,
+                Gender = _dbUser.Gender.ToString(),
                 City = _dbUser.City,
                 LastName = _dbUser.LastName,
-                Status = (UserStatus)_dbUser.Status,
+                Status = _dbUser.Status.ToString(),
                 IsAdmin = _dbUser.IsAdmin,
                 About = _dbUser.About,
                 Rate = _dbUser.Rate,
